Fix ambiguous operator<< for int type
Cast arm_joint_count to int explicitly to resolve ambiguity
between std::ostream::operator<<(int) and RLConfig::operator<<(int)
from Unitree SDK2 ddscxx headers.

+++ src/rl_sar/src/rl_sar/go2x5/comm/go2_x5_ros.cpp
@@ -184,7 +184,7 @@ void RL_Real_Go2X5::HandleArmBridgeStateData(
     {
         std::cout << LOGGER::INFO << "Arm bridge state stream detected: transport="
                   << (this->UseArmBridgeIpc() ? "ipc" : "ros")
-                  << ", dof=" << this->arm_joint_count << std::endl;
+                  << ", dof=" << static_cast<int>(this->arm_joint_count) << std::endl;
     }
     else if (result.warn_shadow_only)
     {
